Alpha (0.3)
- Добавлена проверка на наличие пользователя в БД после команды /start.
- Изменён тип БД с INT на BIGINT, для корректного хранения user-id пользователя.
- Была обновлена модель БД.
- Добавлена команда /help, выводящая пользователю доступные команды.
- Обновлены сообщения бота.
- Добавлен пользовательский интерфейс.
- Добавлен метод форматирования погоды.

+++ Weather bot/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import Column, Integer, BigInteger, String
 from sqlalchemy.ext.declarative import declarative_base
 Base = declarative_base()
 
@@ -6,6 +6,6 @@ class User(Base):
     __tablename__ = 'users'
 
     id = Column(Integer, primary_key=True, autoincrement=True)
-    user_id = Column(Integer, unique=True, nullable=False)
+    user_id = Column(BigInteger, unique=True, nullable=False) 
     username = Column(String(50), unique=True, nullable=True)
     preferred_city = Column(String(100), nullable=True)
